pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csa-docs-tools"
version = "1.0.0"
description = "Testing and validation tools for CSA-in-a-Box documentation"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "CSA Documentation Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.9"
dependencies = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "PyYAML>=6.0.1",
    "aiohttp>=3.8.0",
    "Pillow>=10.0.0",
    "markdown-it-py>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "aioresponses>=0.7.4",
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]
all = ["csa-docs-tools[test,dev]"]

[project.urls]
Homepage = "https://github.com/fgarofalo56/csa-inabox-docs"
Repository = "https://github.com/fgarofalo56/csa-inabox-docs.git"
Documentation = "https://fgarofalo56.github.io/csa-inabox-docs"
"Bug Tracker" = "https://github.com/fgarofalo56/csa-inabox-docs/issues"

[project.scripts]
csa-docs-validate = "csa_docs_tools.cli:sync_main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["csa_docs_tools*"]

[tool.setuptools.package-data]
csa_docs_tools = ["py.typed"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=src/csa_docs_tools",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=70",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests", 
    "network: marks tests that require network access",
    "build: marks tests that test documentation building",
    "quality: marks tests that check documentation quality",
    "validation: marks tests that validate documentation structure",
]
filterwarnings = [
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/csa_docs_tools"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = false
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# Ruff configuration
[tool.ruff]
target-version = "py39"
line-length = 120
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL",  # pylint
    "TRY", # tryceratops
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line-too-long (handled by formatter)
    "PLR0913", # too-many-arguments
    "PLR0912", # too-many-branches
    "TRY003",  # raise-vanilla-args
    "B008",    # function-call-in-default-argument
]
unfixable = [
    "ERA001", # commented-out-code
]

[tool.ruff.per-file-ignores]
"tests/**/*" = [
    "PLR2004", # magic-value-comparison
    "ARG001",  # unused-function-argument
    "ARG002",  # unused-method-argument
    "S101",    # assert
    "S106",    # hardcoded-password-func-arg
]

[tool.ruff.isort]
force-sort-within-sections = true
known-first-party = ["csa_docs_tools"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# MyPy configuration
[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

# Note: Black and isort removed — ruff handles formatting and import sorting

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", "build", "dist"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]